Makes structural changes to system, adds in new navbar component
